Extract results to separate components
